+++ src/components/Animation/Train.tsx
@@ -1,0 +1,13 @@
+import Lottie from 'lottie-react';
+
+import train from '@/assets/lottie/train.json';
+
+const style = {
+  width: '350px',
+};
+
+function TrainAnimation() {
+  return <Lottie animationData={train} style={style} />;
+}
+
+export default TrainAnimation;
