+++ src/constants/member.ts
@@ -1,0 +1,7 @@
+const ROLE = {
+  USER: 'USER',
+  AGENT: 'AGENT',
+  ADMIN: 'ADMIN',
+};
+
+export { ROLE };
